# contracts/openapi/ascendany-v2.yaml
openapi: 3.1.0
info:
  title: AscendAny v2 API
  version: 0.22.0
  description: Authoritative public contract for the Go-only AscendAny v2 runtime.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v2/capabilities:
    get:
      operationId: getCapabilities
      summary: Read active upload contracts and hard limits.
      security: []
      responses:
        "200":
          description: Exact upload contract and active hard limits.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Host-ascendany_refresh
      description: >-
        Opaque rotating refresh credential. The server sets this cookie with
        HttpOnly; Secure; SameSite=None; Path=/ and enforces an exact Origin/CORS
        allowlist. The cookie has no Domain attribute and JavaScript never
        receives its value.
...
  responses:
...
    BadRequest:
      description: Request syntax or payload contract is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
    Forbidden:
      description: Authorization, Origin, or CSRF policy rejected the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
    TooManyRequests:
      description: The operation exceeded its authentication/request rate limit or active SSE capacity.
      headers:
        Retry-After:
          description: Delay in seconds before the caller may retry.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
  schemas:
...
    Capabilities:
      type: object
      additionalProperties: false
      required:
        - pintiaSnapshotSchema
        - pintiaSchemaSha256
        - maxUploadBytes
        - maxProblems
        - maxParticipants
        - maxSubmissions
        - maxCodeBytes
        - writesEnabled
      properties:
        pintiaSnapshotSchema:
          const: ascendany.pintia.snapshot.v2
        pintiaSchemaSha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        maxUploadBytes:
          type: integer
          format: int64
          minimum: 1
        maxProblems:
          type: integer
          minimum: 1
        maxParticipants:
          type: integer
          minimum: 1
        maxSubmissions:
          type: integer
          minimum: 1
        maxCodeBytes:
          type: integer
          minimum: 1
        writesEnabled:
          type: boolean
    CanonicalUUIDv4:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
...
    APIError:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          $ref: "#/components/schemas/CanonicalUUIDv4"
        details:
          type: object
          additionalProperties: true